requirements.txt
tokenizers
click
python-Levenshtein
numpy